remove loader pkg
