+++ .env
@@ -1,0 +1,2 @@
+MNEMONICS="winner giant reward damage expose pulse recipe manual brand volcano dry avoid"
+NETWORK=dev
